# [REVIEW] Avoid unnecessary split for degenerate case where all labels are identical