--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -96,6 +96,7 @@
 - PR #3190: Fix Attribute error on ICPA #3183 and PCA input type
 - PR #3208: Fix EXITCODE override in notebook test script
 - PR #3216: Ignore splits that do not satisfy constraints
+- PR #3243: Avoid unnecessary split for degenerate case where all labels are identical
 
 # cuML 0.16.0 (23 Oct 2020)