--- a/cpp/src/decisiontree/batched-levelalgo/kernels.cuh
+++ b/cpp/src/decisiontree/batched-levelalgo/kernels.cuh
@@ -358,8 +358,6 @@ __global__ void computeSplitRegressionKernel(
   auto* spred = alignPointer<DataT>(smem);
   auto* scount = alignPointer<int>(spred + len);
   auto* sbins = alignPointer<DataT>(scount + nbins);
-
-  // used only for MAE criterion
   auto* spred2 = alignPointer<DataT>(sbins + nbins);
   auto* spred2P = alignPointer<DataT>(spred2 + len);
   auto* spredP = alignPointer<DataT>(spred2P + nbins);
@@ -372,8 +370,6 @@ __global__ void computeSplitRegressionKernel(
   }
   for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
     scount[i] = 0;
-    // printf("indexing from sbins: %p  to %p, sizeof: %d (spred: %p)\n", sbins,
-    //        &sbins[i], (int)sizeof(DataT*), spred);
     sbins[i] = input.quantiles[col * nbins + i];
   }
   __syncthreads();
@@ -403,35 +399,37 @@ __global__ void computeSplitRegressionKernel(
   }
   __threadfence();  // for commit guarantee
   __syncthreads();
-  // for MAE computation, we'd need a 2nd pass over data :(
+
+  /* Make a second pass over the data to compute gain */
+  // Wait until all blockIdx.x's are done
+  MLCommon::GridSync gs(workspace, MLCommon::SyncType::ACROSS_X, false);
+  gs.sync();
+  // now, compute the mean value to be used for metric update
+  for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
+    scount[i] = count[gcOffset + i];
+    spred2P[i] = DataT(0.0);
+  }
+  for (IdxT i = threadIdx.x; i < len; i += blockDim.x) {
+    spred[i] = pred[gOffset + i];
+    spred2[i] = DataT(0.0);
+  }
+  __syncthreads();
+  for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
+    spredP[i] = spred[i] + spred[i + nbins];
+  }
+  __syncthreads();
+  auto invlen = DataT(1.0) / range_len;
+  for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
+    auto cnt_l = DataT(scount[i]);
+    auto cnt_r = DataT(range_len - scount[i]);
+    spred[i] /= cnt_l;
+    spred[i + nbins] /= cnt_r;
+    spredP[i] *= invlen;
+  }
+  __syncthreads();
+
+  // 2nd pass over data to compute partial metric across blockIdx.x's
   if (splitType == CRITERION::MAE) {
-    // wait until all blockIdx.x's are done
-    MLCommon::GridSync gs(workspace, MLCommon::SyncType::ACROSS_X, false);
-    gs.sync();
-    // now, compute the mean value to be used for MAE update
-    for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
-      scount[i] = count[gcOffset + i];
-      spred2P[i] = DataT(0.0);
-    }
-    for (IdxT i = threadIdx.x; i < len; i += blockDim.x) {
-      spred[i] = pred[gOffset + i];
-      spred2[i] = DataT(0.0);
-    }
-    __syncthreads();
-    for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
-      spredP[i] = spred[i] + spred[i + nbins];
-    }
-    __syncthreads();
-    auto invlen = DataT(1.0) / range_len;
-    for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
-      auto cnt_l = DataT(scount[i]);
-      auto cnt_r = DataT(range_len - scount[i]);
-      spred[i] /= cnt_l;
-      spred[i + nbins] /= cnt_r;
-      spredP[i] *= invlen;
-    }
-    __syncthreads();
-    // 2nd pass over data to compute partial MAE's across blockIdx.x's
     for (auto i = range_start + tid; i < end; i += stride) {
       auto row = input.rowids[i];
       auto d = input.data[row + coloffset];
@@ -444,17 +442,31 @@ __global__ void computeSplitRegressionKernel(
         atomicAdd(spred2P + b, raft::myAbs(label - spredP[b]));
       }
     }
-    __syncthreads();
-    // update the corresponding global location
-    for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
-      atomicAdd(pred2P + gcOffset + i, spred2P[i]);
+  } else {
+    for (auto i = range_start + tid; i < end; i += stride) {
+      auto row = input.rowids[i];
+      auto d = input.data[row + coloffset];
+      auto label = input.labels[row];
+      for (IdxT b = 0; b < nbins; ++b) {
+        auto isRight = d > sbins[b];  // no divergence
+        auto offset = isRight * nbins + b;
+        auto diff = label - (isRight ? spred[nbins + b] : spred[b]);
+        auto diff2 = label - spredP[b];
+        atomicAdd(spred2 + offset, (diff * diff));
+        atomicAdd(spred2P + b, (diff2 * diff2));
+      }
     }
-    for (IdxT i = threadIdx.x; i < len; i += blockDim.x) {
-      atomicAdd(pred2 + gOffset + i, spred2[i]);
-    }
-    __threadfence();  // for commit guarantee
-    __syncthreads();
   }
+  __syncthreads();
+  // update the corresponding global location
+  for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
+    atomicAdd(pred2P + gcOffset + i, spred2P[i]);
+  }
+  for (IdxT i = threadIdx.x; i < len; i += blockDim.x) {
+    atomicAdd(pred2 + gOffset + i, spred2[i]);
+  }
+  __threadfence();  // for commit guarantee
+  __syncthreads();
   // last threadblock will go ahead and compute the best split
   bool last = true;
   if (gridDim.x > 1) {
@@ -466,27 +478,15 @@ __global__ void computeSplitRegressionKernel(
   // last block computes the final gain
   Split<DataT, IdxT> sp;
   sp.init();
-  if (splitType == CRITERION::MSE) {
-    for (IdxT i = threadIdx.x; i < len; i += blockDim.x) {
-      spred[i] = pred[gOffset + i];
-    }
-    for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
-      scount[i] = count[gcOffset + i];
-    }
-    __syncthreads();
-    mseGain(spred, scount, sbins, sp, col, range_len, nbins, min_samples_leaf,
-            min_impurity_decrease);
-  } else {
-    for (IdxT i = threadIdx.x; i < len; i += blockDim.x) {
-      spred2[i] = pred2[gOffset + i];
-    }
-    for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
-      spred2P[i] = pred2P[gcOffset + i];
-    }
-    __syncthreads();
-    maeGain(spred2, spred2P, scount, sbins, sp, col, range_len, nbins,
-            min_samples_leaf, min_impurity_decrease);
+  for (IdxT i = threadIdx.x; i < len; i += blockDim.x) {
+    spred2[i] = pred2[gOffset + i];
   }
+  for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
+    spred2P[i] = pred2P[gcOffset + i];
+  }
+  __syncthreads();
+  regressionMetricGain(spred2, spred2P, scount, sbins, sp, col, range_len,
+                       nbins, min_samples_leaf, min_impurity_decrease);
   __syncthreads();
   sp.evalBestSplit(smem, splits + nid, mutex + nid);
 }