--- a/cpp/src/decisiontree/batched-levelalgo/metrics.cuh
+++ b/cpp/src/decisiontree/batched-levelalgo/metrics.cuh
@@ -185,62 +185,7 @@ DI void entropyGain(int* shist, DataT* sbins, Split<DataT, IdxT>& sp, IdxT col,
 }
 
 /**
- * @brief Compute gain based on MSE
- *
- * @param[in]    spred                 left/right child mean prediction for all
- *                                     bins [dim = 2 x bins]
- * @param[in]    scount                left child count for all bins
- *                                     [len = nbins]
- * @param[in]    sbins                 quantiles for the current column
- *                                     [len = nbins]
- * @param[inout] sp                    will contain the per-thread best split
- *                                     so far
- * @param[in]    col                   current column
- * @param[in]    len                   total number of samples for the current
- *                                     node to be split
- * @param[in]    nbins                 number of bins
- * @param[in]    min_samples_leaf      minimum number of samples per each leaf.
- *                                     Any splits that lead to a leaf node with
- *                                     samples fewer than min_samples_leaf will
- *                                     be ignored.
- * @param[in]    min_impurity_decrease minimum improvement in MSE metric. Any
- *                                     splits that do not improve (decrease)
- *                                     the MSE metric at least by this amount
- *                                     will be ignored.
- */
-template <typename DataT, typename IdxT>
-DI void mseGain(DataT* spred, IdxT* scount, DataT* sbins,
-                Split<DataT, IdxT>& sp, IdxT col, IdxT len, IdxT nbins,
-                IdxT min_samples_leaf, DataT min_impurity_decrease) {
-  auto invlen = DataT(1.0) / len;
-  for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
-    auto nLeft = scount[i];
-    auto nRight = len - nLeft;
-    DataT gain;
-    // if there aren't enough samples in this split, don't bother!
-    if (nLeft < min_samples_leaf || nRight < min_samples_leaf) {
-      gain = -NumericLimits<DataT>::kMax;
-    } else {
-      auto invLeft = DataT(1.0) / nLeft;
-      auto invRight = DataT(1.0) / nRight;
-      auto valL = spred[i];
-      auto valR = spred[nbins + i];
-      // parent sum is basically sum of its left and right children
-      auto valP = (valL + valR) * invlen;
-      gain = -valP * valP;
-      gain += valL * invlen * valL * invLeft;
-      gain += valR * invlen * valR * invRight;
-    }
-    // if the gain is not "enough", don't bother!
-    if (gain <= min_impurity_decrease) {
-      gain = -NumericLimits<DataT>::kMax;
-    }
-    sp.update({sbins[i], col, gain, nLeft});
-  }
-}
-
-/**
- * @brief Compute gain based on MAE
+ * @brief Compute gain based on MSE or MAE
  *
  * @param[in]    spred                 left/right child sum of abs diff of
  *                                     prediction for all bins [dim = 2 x bins]
@@ -266,9 +211,10 @@ DI void mseGain(DataT* spred, IdxT* scount, DataT* sbins,
  *                                     will be ignored.
  */
 template <typename DataT, typename IdxT>
-DI void maeGain(DataT* spred, DataT* spredP, IdxT* scount, DataT* sbins,
-                Split<DataT, IdxT>& sp, IdxT col, IdxT len, IdxT nbins,
-                IdxT min_samples_leaf, DataT min_impurity_decrease) {
+DI void regressionMetricGain(DataT* spred, DataT* spredP, IdxT* scount,
+                             DataT* sbins, Split<DataT, IdxT>& sp, IdxT col,
+                             IdxT len, IdxT nbins, IdxT min_samples_leaf,
+                             DataT min_impurity_decrease) {
   auto invlen = DataT(1.0) / len;
   for (IdxT i = threadIdx.x; i < nbins; i += blockDim.x) {
     auto nLeft = scount[i];
@@ -278,9 +224,7 @@ DI void maeGain(DataT* spred, DataT* spredP, IdxT* scount, DataT* sbins,
     if (nLeft < min_samples_leaf || nRight < min_samples_leaf) {
       gain = -NumericLimits<DataT>::kMax;
     } else {
-      gain = spredP[i];
-      gain -= spred[i];
-      gain -= spred[i + nbins];
+      gain = spredP[i] - spred[i] - spred[i + nbins];
       gain *= invlen;
     }
     // if the gain is not "enough", don't bother!